server_py/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
python-dotenv==1.0.1

# HTTP client
httpx==0.27.2

# SSE support
sse-starlette==2.1.3

# Database
pymongo==4.10.1
psycopg2-binary==2.9.10

# Document processing
python-docx==1.1.2
PyPDF2==3.0.1

# File handling
aiofiles==24.1.0
